+++ lib/format_label.dart
@@ -4,9 +4,10 @@ import 'package:flutter/material.dart';
 /// [size], [maxLines] and [textAlign] are optional properties
 class FormatLabel extends StatelessWidget {
   final String mainText;
-  final String text;
-  final Color mainColor;
-  final Color descriptionColor;
+  final String description;
+  final Color textColor;
+  final FontWeight mainTextWeight;
+  final FontWeight? descriptionTextWeight;
   final double? size;
   final int? maxLines;
   final TextAlign? textAlign;
@@ -15,9 +16,10 @@ class FormatLabel extends StatelessWidget {
   const FormatLabel({
     Key? key,
     required this.mainText,
-    required this.text,
-    required this.mainColor,
-    required this.descriptionColor,
+    required this.description,
+    required this.textColor,
+    required this.mainTextWeight,
+    this.descriptionTextWeight = FontWeight.normal,
     this.size = 14,
     this.maxLines = 1,
     this.textAlign = TextAlign.left,
@@ -30,15 +32,15 @@ class FormatLabel extends StatelessWidget {
         maxLines: maxLines,
         text: TextSpan(
             style: TextStyle(
-                fontSize: size, fontWeight: FontWeight.bold, color: mainColor),
+                fontSize: size, color: textColor, fontWeight: mainTextWeight),
             text: '$mainText: ',
             children: [
               TextSpan(
-                  text: text,
+                  text: description,
                   style: TextStyle(
                       fontSize: size,
-                      fontWeight: FontWeight.normal,
-                      color: descriptionColor))
+                      fontWeight: descriptionTextWeight,
+                      color: textColor))
             ]));
   }
 }