+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+* Generalize text color.
+* Fix description text name.
+* Allow to choose the main font weight text style.
+
 ## 0.0.2
 * Text formatted.
 * Documentation.